requirements.txt
asgiref==3.9.1
charset-normalizer==3.4.2
Django==4.2.23
django-widget-tweaks==1.5.0
et_xmlfile==2.0.0
numpy==2.0.2
openpyxl==3.1.5
pandas==2.3.1
pillow==11.3.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
reportlab==4.4.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
Django==4.2.23
gunicorn
whitenoise
dj-database-url